+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 - `limestone/ThemeDecorator/screenTypes` value `wuhd` to support WUHD displays
 
+### Fixed
+
+- `limestone/ContextualPopupDecorator` to be positioned appropriately on dynamic resolution changes
+
 ## [1.6.2] - 2025-11-05
 
 ### Fixed